(DInput joypad) Cut down on code duplication
